Remove QueueMessageExtensions
QueueMessageExtensions allowed changing the Id and PopRecipt properties
of an existing `CloudQueueMessage`. This was used in the case where we
are copying a posined message into the posion queue. The older SDK would
overwrite the exsting state of the message object, which is no longer
possible with the newer SDK.

Because of this, we can simply remove the method and any code that calls
it.