Move `QueueListener.ExecuteAsync` to newer SDK
There's a few things going on here:

- It is no longer possible to control the ClientRequestId on an
operational level.  Instead, the ClientRequestId is set via middlewear
on the HttpPipleline which is something that's done at the client level.
It may be possible to control to have finer grained control over the
ClientRequestId by using a custom Pipleine, but for now, I've just
removed the request ID tracking, as it was already used in logging.

- `CloudQueue.ExistsAsync` did not exist, so I've added an extension
method for it an opened an upstream issue to see if it can be added back
to the SDK at a later date.