Add AsString and AsBytes extensions to QueueMessage
The implementations of these methods are not 100% correct in general
(the V11 SDK had configurable support for doing Base64 encoding and
decoding when moving to and from the wire).  Adding these extension
methods for now to get more code building again and still have an
understanding of the semantics the existing code wanted.