Do not use `QueueMessage` as an "input" type.
In the v11 SDK, if you wanted to send a message, you would create a
`CloudQueueMessage` object and then pass it to a sending method on
`CloudQueue`.  `CloudQueueMessage` was also the type you'd get back when
fetching messages from a Queue.

In the newer SDK, `QueueMessage` is only an "output" type, you get them
by getting items from the queue and the can't usually be constructed.

In addition, in the older SDK, `CloudQueueMessage` would track if it was
created from bytes or a string and that could play into how the message
was sent.

This change introduces a new marker type `QueueMessageInput` which
basically wraps a string and is used when sending a message. It's
helpful if we need to go back and do some of this Base 64 stuff later or
we decide that it's helpful to have a richer type than just "string"
floating around for outputs.

As part of this, I got very confused about what
`QueueTriggerBinding!CreateConverter` was doing, it seemed like it
provided a way to bind against a string and treat it as a QueueMessage.
We don't really support that anymore and it's unclear in this world what
the heck would even happen for binding later (e.g. if you create a
QueueMessage from a string, there's no `MessageId`, what happens if you
try to bind against that later).

I've removed this for now but there's perhaps we can bring it back if
this had a semantics in the past that was reasonable. For the time
being, it was not obvious what a translation would look like.